Renamed package again. Merge reset it. Main change was that I wasn't updating the environment description when checking for the versions.
