+++ src/main/java/br/com/ingenieux/jenkins/plugins/awsebdeployment/SlaveDeployerCallable.java
@@ -14,12 +14,12 @@
  *    limitations under the License.
  */
 
-package br.com.twistedst.jenkins.plugins.awsebdeployment;
+package br.com.ingenieux.jenkins.plugins.awsebdeployment;
 
 import java.io.Serializable;
 
-import br.com.twistedst.jenkins.plugins.awsebdeployment.cmd.DeployerChain;
-import br.com.twistedst.jenkins.plugins.awsebdeployment.cmd.DeployerContext;
+import br.com.ingenieux.jenkins.plugins.awsebdeployment.cmd.DeployerChain;
+import br.com.ingenieux.jenkins.plugins.awsebdeployment.cmd.DeployerContext;
 import jenkins.security.MasterToSlaveCallable;
 
 public class SlaveDeployerCallable extends MasterToSlaveCallable<Boolean, Exception>
