+++ src/main/java/br/com/ingenieux/jenkins/plugins/awsebdeployment/cmd/DeployerChain.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package br.com.twistedst.jenkins.plugins.awsebdeployment.cmd;
+package br.com.ingenieux.jenkins.plugins.awsebdeployment.cmd;
 
 import com.google.common.collect.Lists;
